modularizar queries y mutations  de usuarios
