+++ db/resolvers.js
@@ -1,8 +1,6 @@
 require('dotenv').config({ path:'variables.env' });
 const { GraphQLScalarType } = require('graphql');
 const { Kind } = require('graphql/language');
-const bcryptjs = require('bcryptjs');
-const jwt = require('jsonwebtoken');
 
 const PG = require('../models/PG');
 const CPE = require('../models/CPE');
@@ -14,7 +12,6 @@ const CPG = require('../models/CPG');
 const Insumo = require('../models/Insumos');
 const Pedido = require('../models/Pedidos');
 const Salida = require('../models/Salidas');
-const Usuario = require('../models/Usuarios');
 const Cliente = require('../models/Clientes');
 const Ingreso = require('../models/Ingresos');
 const Producto = require('../models/Productos');
@@ -22,11 +19,11 @@ const StockInsumo = require('../models/StockInsumos');
 const StockInsumos = require('../models/StockInsumos');
 const StockProducto = require('../models/StockProductos');
 
-const crearToken = (usuario, secreta, expiresIn) => {
-    //console.log(usuario);
-    const { id, email, nombre, apellido, rol } = usuario;
-    return jwt.sign( {id, email, nombre, apellido, rol }, secreta)
-}
+// IMPORT  QUERIES
+const querieUsuario = require('./queries/usuarios');
+
+// IMPORT MUTATIONS
+const mutationUsuario = require('./mutations/usuarios');
 
 //RESOLVERS
 const resolvers = {
@@ -49,19 +46,8 @@ const resolvers = {
     }),
 
     Query: {
-        
-        obtenerUsuario: async (_, { }, ctx) => {
-            return ctx.usuario;
-        },
-
-        obtenerUsuarios: async () => {
-            try {
-                const usuarios = await Usuario.find({});
-                return usuarios
-            } catch (error) {
-                console.log(error);
-            }
-        },
+        obtenerUsuario: querieUsuario.obtener,
+        obtenerUsuarios: querieUsuario.obtenerTodos,
 
         obtenerProductos: async () => {
             try {
@@ -232,7 +218,7 @@ const resolvers = {
         },
 
         obtenerProductosTerminados: async () => {
-            let listaProductos = await Producto.find({});
+            let listaProductos = await Producto.find();
             let stockProductos = await StockProducto.find({});
 
             let productos = [];
@@ -454,124 +440,6 @@ const resolvers = {
             return insumosPorProducto;
         },
 
-        obtenerCliente: async (_, {id}) => {
-            //Verificar existencia
-            const cliente = await Cliente.findById(id);
-
-            if (!cliente) {
-                throw new Error('Cliente no encontrado');
-            }
-
-            return cliente;
-        },
-
-        obtenerClientes: async () => {
-            try {
-                const clientes = await Cliente.find({});
-                return clientes;
-            } catch (error) {
-                console.log(error);
-            }
-        },
-
-        obtenerClientesVendedor: async (_, {}, ctx ) => {
-            try {
-                if (ctx.usuario) {
-                const clientes = await Cliente.find({ vendedor: ctx.usuario.id.toString() });
-                return clientes;
-                }
-            } catch (error) {
-                console.log(error);
-            }
-        },
-
-        obtenerPedidos: async () => {
-            try {
-                const pedidos = await Pedidos.find({});
-                return pedidos;
-            } catch (error) {
-                console.log(error);
-            }
-        },
-
-        obtenerPedidosVendedor: async (_, {}, ctx) => {
-            try {
-                const pedidos = await Pedido.find({ vendedor: ctx.usuario.id });
-                return pedidos;
-            } catch (error) {
-                console.log(error);
-            }
-        },
-
-        obtenerPedido: async (_, {id}, ctx) => {
-                //Verificar su existencia
-                const pedido = await Pedido.findById(id);
-                if(!pedido) {
-                    throw new Error('Pedido no encontrado');
-                }
-
-                //Solo lo ve su vendedor
-                if(pedido.vendedor.toString() !== ctx.usuario.id){
-                    throw new Error('Acción no permitida');
-                }
-                
-                //Retornar resultado
-                return pedido;
-        },
-
-        obtenerPedidosEstado: async (_, { estado }, ctx) => {
-            const pedidos = await Pedido.find({ vendedor: ctx.usuario.id, estado});
-
-            return pedidos;
-        },
-
-        mejoresClientes: async () => {
-            const clientes = await Pedido.aggregate([
-                { $match: { estado: "COMPLETADO" }},
-                { $group: {
-                    _id: "$cliente",
-                    total: { $sum: "$total"}
-                }},
-                {
-                    $lookup: {
-                        from: 'clientes',
-                        localField: '_id',
-                        foreignField: "_id",
-                        as: "cliente"
-                    }   
-                },
-                {
-                    $sort: { total:  -1}
-                }
-            ])
-        },
-
-        mejoresVendedores: async () => {
-            const vendedores = await Pedido.aggregate([
-                { $match: { estado: "COMPLETADO" }},
-                { $group: {
-                    _id: "vendedor",
-                    total: { $sum: "$total"}
-                }},
-                {
-                    $lookup: {
-                        from: 'usuarios',
-                        localField: '_id',
-                        foreignField: '_id',
-                        as: "vendedor"
-                    }
-                },
-                {
-                    $limit: 3
-                },
-                {
-                    $sort: { total: -1}
-                }
-            ]);
-
-            return vendedores;
-        },
-
         buscarProducto: async (_, { texto }) => {
             const productos = await Producto.find({$text: { $search: texto }})
 
@@ -755,96 +623,10 @@ const resolvers = {
     },
     
     Mutation: {
-
-        nuevoUsuario: async (_, { input }) => {
-
-            const { email, password } = input;
-
-            //Verificar si ya existe el usuario
-            const existeUsuario = await Usuario.findOne({email});
-            if (existeUsuario) {
-                throw new Error('El usuario ya esta registrado');
-            }
-
-            //Codificar password
-            const salt = await bcryptjs.genSalt(10);
-            input.password = await bcryptjs.hash(password, salt);
-
-            //Guardar en la base de datos
-            try {
-                const usuario = new Usuario(input);
-                usuario.save();
-                return usuario;
-            } catch (error) {
-                console.log(error);
-            }
-        },
-
-        modificarPassword: async (_, {id, input}) => {
-            const {password, newPassword } = input;
-
-            const usuario = await Usuario.findById(id);
-
-            //Comprobar validez de password
-            const passwordCorrecto = await bcryptjs.compare( password, usuario.password );
-            if (!passwordCorrecto) {
-                throw new Error('Contraseña incorrecta');
-            }
-
-            //Codificar nuevo password
-            const salt = await bcryptjs.genSalt(10);
-            nuevoPassword = await bcryptjs.hash(newPassword, salt);
-
-            const resultado = await Usuario.findByIdAndUpdate({_id: id}, {password: nuevoPassword}, {new: true});
-
-            if (resultado) {
-                return 'Contraseña actualizada con exito.'
-            } else {
-                return 'Hubo un error al actualizar la contraseña.'
-            }
-
-        },
-
-        autenticarUsuario: async (_, {input}) => {
-
-            const { password, nombre } = input;
-            const existeUsuario = await Usuario.findOne({nombre});    
-
-            //Verificar si ya existe el usuario
-            if (!existeUsuario) {
-                throw new Error('Usuario incorrecto');
-            }            
-            //Comprobar validez de password
-            const passwordCorrecto = await bcryptjs.compare( password, existeUsuario.password );
-            if (!passwordCorrecto) {
-                throw new Error('Contraseña incorrecta');
-            }
-
-            //Generar Token
-            return {
-                token: crearToken(existeUsuario, process.env.SECRETA, '24h' ),
-            }
-        },
-
-        actualizarUsuario: async (_, {id, input}) => {
-            const { password, nombre, apellido, email } = input;
-            let usuario = await Usuario.findById(id);
-
-            //Comprobar validez de password
-            const passwordCorrecto = await bcryptjs.compare( password, usuario.password );
-            if (!passwordCorrecto) {
-                throw new Error('Contraseña incorrecta');
-            }
-
-            // Actualizar Usuario
-            const usuarioActualizado = await Usuario.findOneAndUpdate(
-                {_id: usuario.id}, 
-                {nombre: nombre, apellido: apellido, email: email}, 
-                {new: true}
-            )
-
-            return usuarioActualizado;
-        },
+        nuevoUsuario: mutationUsuario.crear,
+        autenticarUsuario: mutationUsuario.autenticar,
+        actualizarUsuario: mutationUsuario.actualizar,
+        modificarPassword: mutationUsuario.modificarPassword,
 
         nuevoProducto: async (_, {input}) => {
 
@@ -1048,164 +830,6 @@ const resolvers = {
 
             return "Lote eliminado del stock.";
 
-        },
-
-        nuevoCliente: async (_, { input }, ctx) => {
-
-            //Verificar si ya existe el cliente
-            const { email } = input
-            const cliente = await Cliente.findOne({ email });
-            if(cliente) {
-                throw new Error('Ya existe el cliente');
-            }
-
-            const nuevoCliente = new Cliente(input);
-
-            //Asignar el vendedor
-            nuevoCliente.vendedor = ctx.usuario.id;
-
-            //Guardar en DB
-            try {
-                const resultado = await nuevoCliente.save();
-                return resultado; 
-            } catch (error) {
-                console.log(error);
-            }            
-        },
-
-        actualizarCliente: async (_, {id, input}, ctx) => {
-            //Verificar existencia
-            let cliente = await Cliente.findById(id);
-            if (!cliente) {
-                throw new Error('El cliente no existe');
-            }
-
-            //Verificar si edita el vendedor
-            if(cliente.vendedor.toString() !== ctx.usuario.id ) {
-                throw new Error('No tienes las credenciales.');
-            } 
-
-            //Guardar en db
-            cliente = await Cliente.findOneAndUpdate({_id: id}, input, {new: true});
-            return cliente;
-        },
-
-        eliminarCliente: async (_, { id }, ctx) => {
-            //Verificar existencia
-            let cliente = await Cliente.findById(id);
-            if (!cliente) {
-                throw new Error('El cliente no existe');
-            }
-            
-            //Verificar si edita el vendedor
-            if(cliente.vendedor.toString() !== ctx.usuario.id) {
-                throw new Error('No tienes las credenciales.');
-            } 
-            
-           //Eliminar el cliente
-           await Cliente.findOneAndDelete({_id: id});
-           return 'Cliente eliminado'; 
-        },
-
-        nuevoPedido: async (_, {input}, ctx) => {
-            
-            const { cliente } = input
-            
-            //Verificar existencia de cliente
-            let existeCliente = await Cliente.findById(cliente);
-            if (!existeCliente) {
-                throw new Error('El cliente no existe');
-            }
-
-            //Verificar si pertenece al vendedor
-            if(existeCliente.vendedor.toString() !== ctx.usuario.id ) {
-                throw new Error('No tienes las credenciales.');
-            }
-
-            //Verificar stock
-            for await ( const articulo of input.pedido ) {
-                const { id } = articulo;
-                  
-                const producto = await Producto.findById(id);
-
-                if(articulo.cantidad > producto.cantidad) {
-                    throw new Error(`El articulo: ${producto.nombre} exede la cantidad disponible`);
-                } else {
-                    //Descontar articulos del stock
-                    producto.cantidad = producto.cantidad - articulo.cantidad;
-
-                    await producto.save();
-                } 
-            }
-
-            //Crear nuevo pedido
-            const nuevoPedido = new Pedido(input);
-
-            //Asignar vendedor
-            nuevoPedido.vendedor = ctx.usuario.id;
-
-            //Guardar en DB
-            const resultado = await nuevoPedido.save();
-            return resultado;
-        },
-
-        actualizarPedido: async (_,{id, input}, ctx) => {
-            
-            const { cliente } = input;
-
-            //Verificar si existe
-            const existePedido = await Pedido.findById(id);
-            if(!existePedido) {
-                throw new Error('Pedido no encontrado');
-            }
-
-            //Verificar cliente
-            const existeCliente = await Cliente.findById(cliente);
-            if(!existeCliente) {
-                throw new Error('Cliente no encontrado');
-            }
-
-            //Verificar vendedor
-            if(existeCliente.vendedor.toString() !== ctx.usuario.id) {
-                throw new Error('No tienes las credenciales.');
-            }
-
-            //Revisar el stock
-            if (input.pedido) {
-                for await ( const articulo of input.pedido ) {
-                    const { id } = articulo;
-                    
-                    const producto = await Producto.findById(id);
-
-                    if(articulo.cantidad > producto.cantidad) {
-                        throw new Error(`El articulo: ${producto.nombre} exede la cantidad disponible`);
-                    } else {
-                        //Descontar articulos del stock
-                        producto.cantidad = producto.cantidad - articulo.cantidad;
-
-                        await producto.save();
-                    } 
-                }    
-            }
-            
-            //Guardar en DB
-            resultado = await Pedido.findOneAndUpdate({_id: id}, input, {new: true});
-            return resultado;
-        },
-
-        eliminarPedido: async (_, {id}, ctx) => {
-            //Verificar si existe
-            let pedido = await Pedido.findById(id);
-            if(!pedido) {
-                throw new Error('Pedido no encontrado');
-            }
-
-            if(pedido.vendedor.toString() !== ctx.usuario.id) {
-                throw new Error('No tienes las credenciales.');
-            } 
-
-            await Pedido.findByIdAndDelete(id);
-            return 'Pedido eliminado'
         },
 
         nuevoRegistroSalida: async (_, {input}) => {
